add SIMD wsmask from @wdvxdr1123
